Event participation

+++ lib/wik/events/external_calendar/sync.ex
@@ -6,6 +6,7 @@ defmodule Wik.Events.ExternalCalendar.Sync do
   require Logger
 
   alias Utils.Values
+  alias Wik.Events.Event
   alias Wik.Events.ExternalCalendar.Fetch
   alias Wik.Events.ExternalCalendarSubscription
   alias Wik.Events.ExternalEvent
@@ -155,12 +156,19 @@ defmodule Wik.Events.ExternalCalendar.Sync do
                )
              end)
              |> Enum.reduce_while(:ok, fn event, :ok ->
-               case Repo.delete(event) do
-                 {:ok, _deleted_event} ->
-                   {:cont, :ok}
+               if linked_external_event?(event) do
+                 case mark_source_missing(event, seen_at) do
+                   {:ok, _event} -> {:cont, :ok}
+                   {:error, error} -> Repo.rollback(error)
+                 end
+               else
+                 case Repo.delete(event) do
+                   {:ok, _deleted_event} ->
+                     {:cont, :ok}
 
-                 {:error, error} ->
-                   Repo.rollback(error)
+                   {:error, error} ->
+                     Repo.rollback(error)
+                 end
                end
              end)
            end) do
@@ -303,8 +311,24 @@ defmodule Wik.Events.ExternalCalendar.Sync do
       location: Values.blank_to_nil(event.location),
       description: Values.blank_to_nil(event.description),
       event_url: Values.blank_to_nil(event.url),
-      calendar_name: Values.blank_to_nil(calendar_name)
+      calendar_name: Values.blank_to_nil(calendar_name),
+      source_missing_at: nil
     }
+  end
+
+  defp linked_external_event?(%ExternalEvent{id: id}) do
+    from(event in Event,
+      where: event.source_external_event_id == type(^id, :binary_id),
+      select: event.id,
+      limit: 1
+    )
+    |> Repo.exists?()
+  end
+
+  defp mark_source_missing(%ExternalEvent{} = event, seen_at) do
+    event
+    |> Ecto.Changeset.change(status: :cancelled, source_missing_at: seen_at)
+    |> Repo.update()
   end
 
   defp persist_cache(subscription, attrs) do
